requirements.txt
Adafruit-Blinka==6.4.1
adafruit-circuitpython-busdevice==5.0.6
adafruit-circuitpython-rgb-display==3.10.6
Adafruit-DHT==1.4.0
Adafruit-PlatformDetect==3.4.1
Adafruit-PureIO==1.1.8
gpiozero==1.5.1
matplotlib==3.3.4
mh-z19==3.0.1
numpy==1.16.2
pigpio==1.44
Pillow==8.1.2
ping3==2.7.0
requests==2.21.0
requests-oauthlib==1.0.0
responses==0.9.0
rpi-ws281x==4.2.6
RPi.GPIO==0.7.0
ST7789==0.0.2
